Move feed service, getTimeline, vote table into app-view
