+++ packages/pds/src/app-view/api/app/bsky/feed/getTimeline.ts
@@ -1,10 +1,11 @@
 import { InvalidRequestError } from '@atproto/xrpc-server'
-import { Server } from '../../../../lexicon'
+import { Server } from '../../../../../lexicon'
 import { FeedAlgorithm, FeedKeyset, composeFeed } from '../util/feed'
-import { paginate } from '../../../../db/pagination'
-import AppContext from '../../../../context'
+import { paginate } from '../../../../../db/pagination'
+import AppContext from '../../../../../context'
 import { FeedRow } from '../../../../services/feed'
 
+// @TODO getTimeline() will be replaced by composeTimeline() in the app-view
 export default function (server: Server, ctx: AppContext) {
   server.app.bsky.feed.getTimeline({
     auth: ctx.accessVerifier,
@@ -18,7 +19,7 @@ export default function (server: Server, ctx: AppContext) {
         throw new InvalidRequestError(`Unsupported algorithm: ${algorithm}`)
       }
 
-      const feedService = ctx.services.feed(ctx.db)
+      const feedService = ctx.services.appView.feed(ctx.db)
 
       const followingIdsSubquery = db
         .selectFrom('follow')
