+++ packages/pds/src/app-view/api/app/bsky/util/feed.ts
@@ -1,6 +1,6 @@
 import { AtUri } from '@atproto/uri'
-import { TimeCidKeyset } from '../../../../db/pagination'
-import { Main as FeedViewPost } from '../../../../lexicon/types/app/bsky/feed/feedViewPost'
+import { TimeCidKeyset } from '../../../../../db/pagination'
+import { Main as FeedViewPost } from '../../../../../lexicon/types/app/bsky/feed/feedViewPost'
 import { FeedRow, FeedService } from '../../../../services/feed'
 
 // Present post and repost results into FeedViewPosts
