+++ packages/pds/tests/account-deletion.test.ts
@@ -11,15 +11,15 @@ import { CID } from 'multiformats/cid'
 import { RepoRoot } from '../src/db/tables/repo-root'
 import { UserAccount } from '../src/db/tables/user-account'
 import { IpldBlock } from '../src/db/tables/ipld-block'
-import { Post } from '../src/db/tables/post'
-import { Vote } from '../src/db/tables/vote'
-import { Repost } from '../src/db/tables/repost'
-import { Follow } from '../src/db/tables/follow'
+import { Post } from '../src/app-view/db/tables/post'
+import { Vote } from '../src/app-view/db/tables/vote'
+import { Repost } from '../src/app-view/db/tables/repost'
+import { Follow } from '../src/app-view/db/tables/follow'
 import { RepoBlob } from '../src/db/tables/repo-blob'
 import { Blob } from '../src/db/tables/blob'
-import { PostEntity } from '../src/db/tables/post-entity'
-import { PostEmbedImage } from '../src/db/tables/post-embed-image'
-import { PostEmbedExternal } from '../src/db/tables/post-embed-external'
+import { PostEntity } from '../src/app-view/db/tables/post-entity'
+import { PostEmbedImage } from '../src/app-view/db/tables/post-embed-image'
+import { PostEmbedExternal } from '../src/app-view/db/tables/post-embed-external'
 import { RepoCommitHistory } from '../src/db/tables/repo-commit-history'
 import { RepoCommitBlock } from '../src/db/tables/repo-commit-block'
 import { Record } from '../src/db/tables/record'
